apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "bitsandbytes>=0.49.2",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "channels-valkey>=0.3.0",
    "daphne>=4.2.1",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-mongodb-backend>=6.0.2",
    "django-pydantic-field>=0.5.4",
    "django-tasks-db>=0.12.0",
    "django-unfold>=0.83.1",
    "djangorestframework-simplejwt>=5.5.1",
    "docling>=2.82.0",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2026.3.1",
    "honcho>=2.0.0",
    "langchain>=1.2.13",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "psycopg2-binary>=2.9.11",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.3.0",
    "uuid6>=2025.0.1",
]
